% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_codes.R
\name{country_codes}
\alias{country_codes}
\title{Re-codes country names according to country_dict}
\usage{
country_codes(dataset, source)
}
\arguments{
\item{dataset}{A dataframe of one of the package's dataset}

\item{source}{String of dataset, to be found in var_dict}
}
\value{
Dataframe with added country variables
}
\description{
Re-codes country names according to country_dict
}
\examples{
country_codes(df, "ches")
}
