% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cses_polarisation_index.R
\name{cses_polarisation_index}
\alias{cses_polarisation_index}
\title{CSES Polarisation Index}
\usage{
cses_polarisation_index(dataset)
}
\arguments{
\item{dataset}{}
}
\value{
Country/Year dataset with CSES Polar. Index
}
\description{
CSES Polarisation Index
}
