% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Calculate weighted and unweighted Distance measure}
\usage{
distance(dataset)
}
\arguments{
\item{dataset}{Imported Dataset with adapted Varnames, originally CSES IMD}
}
\value{
Dataset with added variables distance and distance_wgt
}
\description{
Calculate weighted and unweighted Distance measure
}
