% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_range.R
\name{ind_range}
\alias{ind_range}
\title{Individual Range}
\usage{
ind_range(data, measure)
}
\arguments{
\item{data}{Dataframe}

\item{measure}{Individual level variable without party suffix}
}
\value{
Dataset wih individual level range score
}
\description{
Individual Range
}
\examples{
cses_5 <- range(cses_5, "leftright")
}
