% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_aggregate.R
\name{mean_aggregate}
\alias{mean_aggregate}
\title{Aggregate mean of polarization measures on a Country-Year Level}
\usage{
mean_aggregate(data, polarization_var)
}
\arguments{
\item{data}{Dataset with respondents as observation units}

\item{polarization_var}{Variable with individual polarization scores}

\item{new_var_name}{Name of the new variable}
}
\value{
Dataset with mean polarization scores for Country-Year units
}
\description{
Aggregate mean of polarization measures on a Country-Year Level
}
