% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{range}
\alias{range}
\title{Range}
\usage{
range(data, measure)
}
\arguments{
\item{data}{Dataframe}

\item{measure}{Individual level variable without party suffix}
}
\value{
Dataset wih individual level range score
}
\description{
Range
}
\examples{
cses_5 <- range(cses_5, "leftright")
}
