% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_missings.R
\name{recode_missings}
\alias{recode_missings}
\title{Recode missings to NA on project-relevant variables in a dataset. Function
creates list of variables itself using global 'var_dict'}
\usage{
recode_missings(dataset, source)
}
\arguments{
\item{dataset}{.df object}
}
\value{
DF with recoded missings
}
\description{
Recode missings to NA on project-relevant variables in a dataset. Function
creates list of variables itself using global 'var_dict'
}
