% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_variables.R
\name{rename_variables}
\alias{rename_variables}
\title{Rename Variables
Renames selected variables in a dataframe to correspond to the package convention, see var_dict}
\usage{
rename_variables(dataset, source)
}
\arguments{
\item{dataset}{A dataframe of one of the package's dataset}

\item{source}{String of dataset, to be found in var_dict}
}
\value{
dataframe with renamed variables
}
\description{
Rename Variables
Renames selected variables in a dataframe to correspond to the package convention, see var_dict
}
\examples{
rename_variables(df, "cses_5")
}
