% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_voteshare.R
\name{replace_voteshare}
\alias{replace_voteshare}
\title{Replace party vote share with president vote share where relevant}
\usage{
replace_voteshare(dataset)
}
\arguments{
\item{dataset}{}
}
\value{
dataset with adapted voteshare measure
}
\description{
Replace party vote share with president vote share where relevant
}
\examples{
cses_5 <- replace_voteshare(cses_5)
}
