% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_aggregate.R
\name{sd_aggregate}
\alias{sd_aggregate}
\title{Aggregate standard deviation of scores on a Country-Year Level}
\usage{
sd_aggregate(data, score_var)
}
\arguments{
\item{data}{Dataset with respondents as observation units}

\item{score_var}{Variable with individual scores to compute SD from}

\item{new_var_name}{Name of the new variable}

\item{country_var}{Country variable in the dataset}

\item{year_var}{Year variable in the dataset}
}
\value{
Dataset with polarization scores for Country-Year units
}
\description{
Aggregate standard deviation of scores on a Country-Year Level
}
