% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_expert_leftright.R
\name{sd_expert_leftright}
\alias{sd_expert_leftright}
\title{Standard deviation of expert party evaluations}
\usage{
sd_expert_leftright(data)
}
\arguments{
\item{data}{Dataset with expert lr expert evaluations as columns}
}
\value{
Dataset with mean polarization scores for Country-Year units
}
\description{
Standard deviation of expert party evaluations
}
