% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_parties.R
\name{sd_parties}
\alias{sd_parties}
\title{SD of Party Perception}
\usage{
sd_parties(dataset, measure)
}
\arguments{
\item{dataset}{Dataframe}

\item{measure}{Individual level variable without party suffix}
}
\value{
Dataset wih individual level sd score
}
\description{
SD of Party Perception
}
\examples{
cses <- sd_parties(cses, "leftright")
}
