% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread.R
\name{spread}
\alias{spread}
\title{Calculate weighted and unweighted Wagner measure}
\usage{
spread(dataset)
}
\arguments{
\item{dataset}{Imported Dataset with adapted Varnames, originally CSES IMD}
}
\value{
Dataset with added variables affpol and affpolwgt (plus helper variables)
}
\description{
Calculate weighted and unweighted Wagner measure
}
