% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagner_measure.R
\name{wagner_measure}
\alias{wagner_measure}
\title{Calculate weighted and unweighted Wagner measure}
\usage{
wagner_measure(dataset)
}
\arguments{
\item{dataset}{Imported Dataset with adapted Varnames, originally CSES IMD}
}
\value{
Dataset with added variables affpol and affpolwgt (plus helper variables)
}
\description{
Calculate weighted and unweighted Wagner measure
}
