% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polaR_import.R
\name{polaR_import}
\alias{polaR_import}
\title{Import polaR datasets
Import the CSES Integrated Dataset.
Takes .dta files, renames and formats package-relevant variables.}
\usage{
polaR_import(source, path, keep_all = TRUE)
}
\arguments{
\item{source}{The source of the data, e.g. CSES 5}

\item{path}{File path string to a .dta file}

\item{keep_all_ess}{defaults to TRUE, keep all variables in the ESS or reduce the dataset size to be able to handle the data}
}
\value{
CSES IMD dataset that is ready for use with other functions
}
\description{
Import polaR datasets
Import the CSES Integrated Dataset.
Takes .dta files, renames and formats package-relevant variables.
}
